pyproject.toml
[project]
name = "ppt"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
  "pytest",
  "pydantic>=2",
  "ruamel.yaml>=0.18",
  "python-dotenv>=1",
]

[tool.pytest.ini_options]
testpaths = ["tests/unit"]
addopts = "-q"
markers = [
  "container: requires container runtime (integration)",
  "gitlab_api: calls GitLab HTTP APIs (integration)",
]
